Update versions of multiple Python packages
Updated the versions of several packages in the poetry.lock file such as certifi, docker, idna, packaging, requests, ruff, types-python-dateutil, types-requests, and urllib3. This update will ensure that the application is running on the latest and most secure versions of these libraries. This is part of regular maintenance and keeping dependencies up to date.